+++ src/Pages/Proposal/Components/index.js
@@ -1,0 +1,1 @@
+export { default as ProposalInformation } from './ProposalInformation';